restructure folders, get rid of visx
