+++ app/package.json
@@ -8,14 +8,9 @@
     "@fortawesome/free-regular-svg-icons": "^5.15.4",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "@fortawesome/react-fontawesome": "^0.1.16",
-    "@visx/curve": "^2.1.0",
-    "@visx/responsive": "^2.4.1",
-    "@visx/text": "^2.3.0",
-    "@visx/xychart": "^2.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-scripts": "^5.0.0",
-    "react-spring": "^9.4.1"
+    "react-scripts": "^5.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
