--- app/src/Chart.css
@@ -1,37 +1,0 @@
-.chart {
-  position: relative;
-  width: calc(var(--page) / 2 - 40px);
-  max-width: 100%;
-  aspect-ratio: 4 / 3;
-  transition: transform var(--fast), box-shadow var(--fast);
-}
-
-.chart:hover {
-  transform: scale(1.01);
-  box-shadow: 0 2px 5px var(--shadow);
-}
-
-.download {
-  position: absolute;
-  left: 0;
-  bottom: 0;
-  width: 40px;
-  height: 40px;
-  background: transparent;
-  color: transparent;
-  transition: color var(--fast);
-}
-
-.chart:hover .download {
-  color: var(--gray);
-}
-
-.chart:hover .download:hover {
-  color: var(--purple);
-}
-
-@media (max-width: 600px) {
-  .chart {
-    box-shadow: none;
-  }
-}
