+++ app/src/sections/Header.js
@@ -1,9 +1,9 @@
 import { useContext } from "react";
-import Title from "./Title";
-import Search from "./Search";
-import { AppContext } from "./App";
+import Title from "../components/Title";
+import Search from "../components/Search";
+import Matrix from "../components/Matrix";
+import { AppContext } from "../App";
 import "./Header.css";
-import Matrix from "./Matrix";
 
 const Header = () => {
   const { fullscreen } = useContext(AppContext);
